backend/pyproject.toml
[project]
name = "personal-website"
version = "0.1.0"
description = ""
authors = [
    {name = "geniungamartya@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-frontmatter (>=1.1.0,<2.0.0)",
    "typer (>=0.15.2,<0.16.0)",
    "minio (>=7.2.15,<8.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "psycopg2[extras] (>=2.9.10,<3.0.0)",
    "bcrypt (==4.0.1)",
    "passlib (>=1.7.4,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pytest = "^8.3.5"
ipdb = "^0.13.13"
mypy = "^1.15.0"

[tool.ruff.lint]
extend-select = ["I"]
